+++ example/android/app/src/main/kotlin/com/dsi/facebook_audience_network_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.dsi.facebook_audience_network_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()