+++ lib/ad/ad_interstitial.dart
@@ -24,7 +24,7 @@ enum InterstitialAdResult {
 class FacebookInterstitialAd {
   static void Function(InterstitialAdResult, dynamic)? _listener;
 
-  static const _channel = const MethodChannel(INTERSTITIAL_AD_CHANNEL);
+  static const _channel = MethodChannel(INTERSTITIAL_AD_CHANNEL);
 
   /// Loads an Interstitial Ad in background. Replace the default [placementId]
   /// with the one which you obtain by signing-up for Facebook Audience Network.